Fase 9 + docs: pruebas minimas, README completo y notebooks
- tests/: preprocesamiento (dataset intacto, fecha valida/unica, lags sin fuga,
  objetivo fuera de predictores, clave unica) y modelos (entrenan, guardan/cargan,
  probabilidades en [0,1], k de clustering, MAPE seguro, periodos de pronostico)
- README.md: seccion "Estado inicial del dataset", instalacion, entrenamiento,
  Supabase, despliegue, metricas reales y "Modificaciones rapidas para la exposicion"
- notebooks/01-04: exploracion y validacion por fase
- Robustez: clamp de k a n_muestras-1 en clustering; edad a float

+++ src/preprocessing.py
@@ -68,7 +68,7 @@ def load_raw(path=None) -> pd.DataFrame:
 
 def _edad_a_anios(edad: pd.Series, tipo: pd.Series) -> pd.Series:
     """Convierte edad a anios usando tipo_edad (A=anios, M=meses, D=dias)."""
-    edad_num = pd.to_numeric(edad, errors="coerce")
+    edad_num = pd.to_numeric(edad, errors="coerce").astype(float)
     tipo = tipo.astype(str).str.upper().str.strip()
     anios = edad_num.copy()
     anios = anios.mask(tipo == "M", edad_num / 12.0)
